+++ src/lokalise/base_client.ts
@@ -22,12 +22,12 @@ export class BaseClient {
 	 * @returns       LokaliseApi object to work with.
 	 */
 	constructor(params: ClientParams) {
-		const apiKey = params["apiKey"];
+		const apiKey = params.apiKey;
 		if (apiKey === null || apiKey === undefined || apiKey.length === 0) {
 			throw new Error("Error: Instantiation failed: Please pass an API key");
 		}
 		this.clientData.token = apiKey;
-		const compression = params["enableCompression"];
+		const compression = params.enableCompression;
 		if (compression !== null && compression !== undefined) {
 			this.clientData.enableCompression = compression;
 		}