+++ src/lokalise/lokalise_api_oauth.ts
@@ -4,7 +4,7 @@ export class LokaliseApiOAuth extends LokaliseApi {
 	constructor(params: ClientParams) {
 		super(params);
 
-		const tokenType = params["tokenType"];
+		const tokenType = params.tokenType;
 		this.clientData.tokenType = tokenType ?? "Bearer";
 
 		this.clientData.authHeader = "Authorization";