+++ src/interfaces/user_group.ts
@@ -2,8 +2,8 @@ export interface UserGroup {
 	group_id: number;
 	name: string;
 	permissions: {
-		is_admin: boolean; //deprecated
-		is_reviewer: boolean; //deprecated
+		is_admin: boolean; // deprecated
+		is_reviewer: boolean; // deprecated
 		admin_rights: string[];
 		languages: Array<{
 			lang_id: number;
@@ -11,11 +11,11 @@ export interface UserGroup {
 			lang_name: string;
 			is_writable: boolean;
 		}>;
-		role_id: number;
 	};
 	created_at: string;
 	created_at_timestamp: number;
 	team_id: number;
 	projects: string[] | number[];
 	members: number[] | string[];
+	role_id: number | null;
 }