+++ src/models/permission_template.ts
@@ -1,7 +1,10 @@
 import type { PermissionTemplate as PermissionTemplateInterface } from "../interfaces/permission_template.js";
 import { BaseModel } from "./base_model.js";
 
-export class PermissionTemplate extends BaseModel implements PermissionTemplateInterface {
+export class PermissionTemplate
+	extends BaseModel
+	implements PermissionTemplateInterface
+{
 	declare id: number;
 	declare role: string;
 	declare permissions: string[];