+++ src/models/user_group.ts
@@ -5,8 +5,8 @@ export class UserGroup extends BaseModel implements UserGroupInterface {
 	declare group_id: number;
 	declare name: string;
 	declare permissions: {
-		is_admin: boolean;//deprecated
-		is_reviewer: boolean;//deprecated
+		is_admin: boolean; // deprecated
+		is_reviewer: boolean; // deprecated
 		admin_rights: string[];
 		languages: Array<{
 			lang_id: number;
@@ -14,11 +14,11 @@ export class UserGroup extends BaseModel implements UserGroupInterface {
 			lang_name: string;
 			is_writable: boolean;
 		}>;
-		role_id: number;
 	};
 	declare created_at: string;
 	declare created_at_timestamp: number;
 	declare team_id: number;
 	declare projects: string[] | number[];
 	declare members: number[] | string[];
+	declare role_id: number | null;
 }