+++ test/user_groups/user_groups.spec.ts
@@ -58,6 +58,8 @@ describe("UserGroups", () => {
 		});
 
 		expect(user_groups.items[0].group_id).to.eq(2500);
+		expect(user_groups.items[0].role_id).to.eq(5);
+		expect(user_groups.items[1].role_id).to.eq(null);
 		expect(user_groups.totalResults).to.eq(4);
 		expect(user_groups.totalPages).to.eq(2);
 		expect(user_groups.resultsPerPage).to.eq(2);
@@ -92,6 +94,7 @@ describe("UserGroups", () => {
 		expect(user_group.team_id).to.eq(teamId);
 		expect(user_group.projects).to.have.lengthOf(0);
 		expect(user_group.members).to.include(34051);
+		expect(user_group.role_id).to.eq(5);
 	});
 
 	it("creates", async () => {