+++ tsup.config.ts
@@ -1,12 +1,12 @@
-import { defineConfig } from 'tsup';
+import { defineConfig } from "tsup";
 
 export default defineConfig({
-  entry: ['src/main.ts'],
-  outDir: 'dist',
-  format: ['esm'],
-  dts: true,
-  splitting: false,
-  shims: true,
-  sourcemap: true,
-  clean: true,
+	entry: ["src/main.ts"],
+	outDir: "dist",
+	format: ["esm"],
+	dts: true,
+	splitting: false,
+	shims: true,
+	sourcemap: true,
+	clean: true,
 });