+++ test/setup.ts
@@ -21,7 +21,7 @@ import type { HttpMethod } from "../src/types/http_method.js";
 
 const mockAgent = new MockAgent();
 setGlobalDispatcher(mockAgent);
-// mockAgent.disableNetConnect();
+mockAgent.disableNetConnect();
 
 type StubParams = {
 	uri: string;